- Changed all parseInt for parseFloat so decimal inputs would be registered - Added validation logic so only numbers could be converted

+++ assets/js/lengthConverter.js
@@ -1,23 +1,23 @@
 export function mToFeet (inputValue) {
-    return parseInt(inputValue) * 3.2808;
+    return parseFloat(inputValue) * 3.2808;
 }
 
 export function mToYard (inputValue) {
-    return parseInt(inputValue) * 1.0936;
+    return parseFloat(inputValue) * 1.0936;
 }
 
 export function ftToMeter (inputValue) {
-    return parseInt(inputValue) * 0.3048;
+    return parseFloat(inputValue) * 0.3048;
 }
 
 export function ftToYard (inputValue) {
-    return parseInt(inputValue) * 0.3333;
+    return parseFloat(inputValue) * 0.3333;
 }
 
 export function ydToMeter (inputValue) {
-    return parseInt(inputValue) * 0.9144;
+    return parseFloat(inputValue) * 0.9144;
 }
 
 export function ydToFeet (inputValue) {
-    return parseInt(inputValue) * 3;
+    return parseFloat(inputValue) * 3;
 }
