+++ assets/js/temperatureConverter.js
@@ -1,23 +1,23 @@
 export function CToFahrenheit (inputValue) {
-    return (parseInt(inputValue) * 9 / 5) + 32;
+    return (parseFloat(inputValue) * 9 / 5) + 32;
 }
 
 export function CToKelvin (inputValue) {
-    return parseInt(inputValue) + 273.15;
+    return parseFloat(inputValue) + 273.15;
 }
 
 export function FToCelcius (inputValue) {
-    return (parseInt(inputValue) - 32) * 5 / 9;
+    return (parseFloat(inputValue) - 32) * 5 / 9;
 }
 
 export function FToKelvin (inputValue) {
-    return (parseInt(inputValue) - 32) * 5 / 9 + 273.15;
+    return (parseFloat(inputValue) - 32) * 5 / 9 + 273.15;
 }
 
 export function KToCelcius (inputValue) {
-    return parseInt(inputValue) - 273.15;
+    return parseFloat(inputValue) - 273.15;
 }
 
 export function KToFahrenheit (inputValue) {
-    return (parseInt(inputValue) - 273.15) * 9 / 5 + 32;
+    return (parseFloat(inputValue) - 273.15) * 9 / 5 + 32;
 }
